Added premature room state checking
Now room state (a.k.a. isFull()) will be checked as soon user joins
the room.

A special variable 'tryLeave' is now added into the socket object to
account for cases where a user has entered a incorrect room ID and the
`disconnect` would try to call `leaveUser()` on an un existant room.

The logging API has been re-written to be cleaner and more case specific

TODO: Add 'forever' support and make the app blow up instead of failing
silently