Added bullets
Each room keeps track of its bullets and update is called at the end
of each turn.

Update calls update on each of the `arena.Bullet' objects (soon they
will be inherited from the entity class). It moves the bullet forward.
It then checks whether the bullet has collied with the player (Right now
it's always `no`). Each bullet has a `goneMark` flag. It's triggered
by hitting the wall, `setGoneMark` (used when checking for collisions)
and if `turnSpanLeft` expires. At the end of the update cycle bullets
are removed from `@bullets` array if `isGone` returns `yes` on a bullet.

'update' event format is now updated too. Each object now contains
'type' filed to judge the type of sprite to render on the client side
and a name. '-bullet-<id>' was chosen for bullets as it's a valid css
class selector and it's not a valid username.

* Comments were added to the regexs in the user routes.
* Default configs were changed.
  - maxMove is now smaller as the default roomSize was always small.
  - Default bullet trunSpan was added
  - Default live for user and hit values were added
* Queue's `jsonify` functions now only sends ship properties that were
  actually modified.
* Socket.io's connect function is now url aganostic
* Fixed typo in templates base