Added live room availability counter
Migrated all the old /public JavaScript to CoffeeScript.

Completely rewrote the way rooms operate and are cached, in order to
add live room availability counter to the /rooms.html. One of the
last touches to general site design before I finalize the arena logic.

The data structures used are now kept in /realtime/storage but the
client logic is still based in /realtime/arena.

There is one central 'rooms' instance that emits events once
the room availability is changed so that /rooms.html clients can be
kept up to date with it.

My previous approach was based on keeping a 'taken' field in the
database.

* Removed "Closed sources" in favour of GPL
* Removed other stuff from package.json
* Added 'stylus' as a dependency so it's not external
* Added an option to completely remove socket.io logging
* '?' is now being used instead of 'isDef' in many places
* A user can now join a room only once
* 'no_room' event is finally renamed 'kick'. It was called that before
  because it was originally only mint to handle when a room didn't
  exist.